fix: linting again
